pmbench Support for powerpc
pmbench was originally developed for the x86 architecture, and as
a result, it contained hardcoded x86-specific instructions to
measure time and perform load/store operations while measuring
read and write latency. Due to this, pmbench was not compiling
correctly on the powerpc architecture.

This patch adds basic support for pmbench on powerpc. With this
patch, pmbench will successfully build and run on powerpc systems
and is now capable of measuring read/write latency. To achieve
this, powerpc specific instructions have been added to read the
time and perform load/store operations for measuring read and
write latency.

This still does not support reading TLB info, cache info and
other statistics.

Signed-off-by: Donet Tom <donettom@linux.ibm.com>